+++ restaurants/urls.py
@@ -1,16 +1,14 @@
 from django.urls import path, include
 from rest_framework.routers import DefaultRouter
 
-from restaurants.views import RestaurantViewset, GuestViewset, get_guest
+from restaurants.views.restaurantView import RestaurantViewset, get_guest
+from restaurants.views.guestView import GuestViewset
 
 router = DefaultRouter(trailing_slash=False)
 router.register(r'restaurant', RestaurantViewset, basename='restaurants')
-router.register('guest', GuestViewset)
+router.register(r'pos', GuestViewset, basename='restaurants')
 
 urlpatterns = [
     path('', include(router.urls)),
-]
-
-urlpatterns += [
-    path('group/', get_guest,),
+    path('group/', get_guest),
 ]
